+++ app/domain_models/follow_and_unfollow_user.rb
@@ -1,18 +1,19 @@
-require_relative "follow_status_data"
+require_relative "../repositories/personal_follow_lists_repository"#personal_follow_lists_repository"
 class FollowAndUnfollowUser
 
-    attr_reader :subject_user_id, :follow_status_data
+    attr_reader :subject_user_id, :follow_list
   
     def initialize(subject_user_account)
       @subject_user_id=User.find_by(account: subject_user_account).id
-      @follow_status_data=FollowStatusData.new()
+      @follow_list=PersonalFollowListsRepository.new(subject_user_account)
     end
 
     def follow(object_user_account)
         object_user_id=User.find_by(account: object_user_account).id
         cannot_follow_yourself!(object_user_id)
         already_following!(object_user_id)
-        follow_status_data.add_follow_status_record(subject_user_id,"follow",object_user_id)
+        succeed_to_create_record("follow",object_user_id)
+
         
     end
 
@@ -20,7 +21,7 @@ class FollowAndUnfollowUser
         object_user_id=User.find_by(account: object_user_account).id
         cannot_follow_yourself!(object_user_id)
         not_following!(object_user_id)
-        follow_status_data.add_follow_status_record(subject_user_id,"unfollow",object_user_id)
+        succeed_to_create_record("unfollow",object_user_id)
     
         
     end
@@ -31,11 +32,15 @@ private
     end
 
     def already_following!(object_user_id)
-        raise AlreadyFollowingError.new(@subject_user_id,object_user_id) if follow_status_data.get_following(@subject_user_id).include?(object_user_id)
+        raise AlreadyFollowingError.new(@subject_user_id,object_user_id) if follow_list.get_following.include?(object_user_id)
     end
 
     def not_following!(object_user_id)
-        raise NotFollowingError.new(@subject_user_id,object_user_id) if not(follow_status_data.get_following(@subject_user_id).include?(object_user_id))
+        raise NotFollowingError.new(@subject_user_id,object_user_id) if not(follow_list.get_following.include?(object_user_id))
+    end
+
+    def succeed_to_create_record(action_type,object_user_id)
+        follow_list.add_follow_status_record(subject_user_id,action_type,object_user_id)
     end
 end
