--- spec/domain_models/follow_status_data_spec.rb
@@ -1,60 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe FollowStatusData do
-    User.create!(account: "zzz",phone_number: "0973006502",email: "eddie@gmail.com",password: "23")
-    User.create!(account: "zzzzz",phone_number: "0973006503",email: "eddie11@gmail.com",password: "23")
-    User.create!(account: "zzzz",phone_number: "0973006504",email: "eddie1@gmail.com",password: "23")
-    follow_status_data=FollowStatusData.new()
-    
-  describe "initialize FollowStatusData with follow_status_record table" do
-
-    it 'check if the test records are properly initialized' do
-       FollowStatusRecord.create!(subject_user_id:1,action_type:"follow",object_user_id:2)
-       FollowStatusRecord.create!(subject_user_id:2,action_type:"follow",object_user_id:3)
-       follow_status_data.update_record_to_latest
-       expect(follow_status_data.get_record_number).to eq(2)
-    end
-
-    it 'check if each_user_following_list and each_user_follower_list are initialized' do
-
-        expect(follow_status_data.each_user_following_list).is_a?(Hash)
-        expect(follow_status_data.each_user_follower_list).is_a?(Hash)
-        
-    end
-
-    it 'check if each_user_following_list is working properly' do
-        
-        expect(follow_status_data.each_user_following_list[1]).to eq([2])
-        expect(follow_status_data.each_user_following_list[2]).to eq([3])
-        
-    end
-
-    it 'check if each_user_follower_list is working properly' do
-        expect(follow_status_data.each_user_follower_list[2]).to eq([1])
-        expect(follow_status_data.each_user_follower_list[3]).to eq([2])
-
-    end
-  end
-
-  describe "get and add record using FollowStatusData" do
-
-    it 'check if get_follower() is working properly' do
-        expect(follow_status_data.get_follower(2)).to eq([1])
-        expect(follow_status_data.get_follower(3)).to eq([2])
-    end
-
-    it 'check if get_following() is working properly' do
-        expect(follow_status_data.get_following(1)).to eq([2])
-        expect(follow_status_data.get_following(2)).to eq([3])
-    end
-
-    it 'check if add_follow_status_record is working properly' do 
-
-        follow_status_data.add_follow_status_record(1,"follow",3)
-        expect(follow_status_data.get_following(1)).to eq([2,3])
-        follow_status_data.add_follow_status_record(1,"unfollow",3)
-        expect(follow_status_data.get_following(1)).to eq([2])
-
-    end
-  end
-end
